refactor: Move formatRelativeTime function to LaunchPrepDetails component for better encapsulation
